+++ SOLIDprinciples/OpenClosed/OpenClosed_GoodDesign/HitPointModifiers/IntimidationHitPointModifier.cs
@@ -2,17 +2,17 @@
 {
     public sealed class IntimidationHitPointModifier : HitPointModifierBase
     {
-        private const double INTIMIDATION_FORCE = 3;
+        private const int INTIMIDATION_FORCE = 3;
 
         public IntimidationHitPointModifier(int modifierValue, int abilityBonus) :
                base(modifierValue, abilityBonus)
         {
         }
 
-        public override double CalculateModifierValue(int hitPointValue)
+        public override int CalculateModifierValue(int hitPointValue)
         {
-            double modifierValue = base.CalculateModifierValue(hitPointValue);
-            double calculatedModifierValue = INTIMIDATION_FORCE + modifierValue;
+            int modifierValue = base.CalculateModifierValue(hitPointValue);
+            int calculatedModifierValue = INTIMIDATION_FORCE + modifierValue;
             return calculatedModifierValue;
         }
     }
